Compilation works basically.
A cabalized install is still not strictly necessary for testing,
which is good. Basically, the current directory the executable is
run from is always included in the module search path.

It still doesn't do the whole execution thing, so that will be next,
followed by the implementation of some special flags.