Added '--force-reconf' and '--dyre-debug' options.
The '--force-reconf' one, quite obviously, forces a recompilation
of the config file, whether we think one is necessary or not.

The '--dyre-debug' option sets the binary dir, config dir, and
temp dir all equal to the current directory. This is suitable
for debugging a program configured using Dyre, when you don't
want to bother cluttering up the rest of your filesystem.

The two options should work together in the obvious ways.