New option in Params to prevent name shadowing.
By default Dyre included current directory in the list of include dirs.
It causes a problem when running configurable program in its own project
directory (particularly natural for an editor.)

This patch adds an option includeCurrentDirectory that stops it from
happening, if set to False. Behaviour is unchanged by default.