compile: remove error file on successful compilation
When compiling we write GHC's standard error output to a file.
Other parts of dyre interpret a non-empty error file as indication
that a compile error occurred.  But there are situations when GHC
writes to standard error, even when the compile succeeds (e.g.
reporting the use of a .ghc.environment file).

To avoid spurious failures in this scenario, after a successful
compile always remove the error file.