tests: merge stdout and stderr
Currently each test's stderr is streamed directly to terminal,
whereas stdout is captured by allTests.sh and displayed only if the
test failed.  This results in confusing output; parts of the test
output appear out of order and in different places.

When executing the test, redirect stderr to stdout so *all* test
output is captured together.