windows: consistent output file naming
The current implementation gets confused about filenames on Windows.
Whereas we tell GHC to output `foo-mingw32-x86_64.exe.tmp`, when we
goes to rename the file we look for `foo-mingw32-x86_64.tmp`, which
does not exist.

Create a function for naming the output file, based on the desired
custom executable name.  Use this function in both places.