Allow to use NIX_GHC next to HC env variable
Nix generates wrapper scripts for packages which ship their own sources
in order to compile them with dyre. Currently they fail to compile,
because dyre can't find the GHC binary (esp. not on NixOS which doesn't
follow the File System Hierarchy Standard).

This patch adds the `NIX_GHC` alternative to pick beside `HC` in order
to find the GHC compiler.

Fixes: https://github.com/willdonnelly/dyre/issues/42